requirements (2).txt
# A2A Multi-Agent Customer Service System
# Assignment 5 - ADSP 32028

# A2A Protocol SDK (Google's official SDK)
a2a-sdk>=0.3.0

# Web Framework
fastapi>=0.115.0
uvicorn>=0.30.0
starlette>=0.38.0
httpx>=0.27.0

# Database
sqlalchemy>=2.0.0

# Utilities
pydantic>=2.0.0

# OpenAI
openai>=1.0.0
